requirements.txt
ultralytics>=8.0.0
opencv-python>=4.8.0
speechrecognition>=3.10.0
pyttsx3>=2.90
googlemaps>=4.10.0
geopy>=2.3.0
pyaudio>=0.2.11
numpy>=1.24.0
pillow>=10.0.0
requests>=2.31.0
python-dotenv>=1.0.0
pywin32>=306
